requirements.txt
plotly==6.5.0
numpy
pandas
openpyxl
st_pages
streamlit-extras